Creation of render pipelines now occurs in pipeline.rs, cleaned up render_context.rs and consolidated duplicate pipelines
